+++ NotSnap/app/src/main/java/com/COMP3004/NotSnap/Camera.java
@@ -1,0 +1,110 @@
+package com.COMP3004.NotSnap;
+
+import android.graphics.Bitmap;
+import android.graphics.Matrix;
+import android.os.Handler;
+import android.os.HandlerThread;
+import android.util.Rational;
+import android.util.Size;
+import android.view.Surface;
+import android.view.TextureView;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.camera.core.CameraX;
+import androidx.camera.core.ImageAnalysis;
+import androidx.camera.core.ImageAnalysisConfig;
+import androidx.camera.core.ImageCapture;
+import androidx.camera.core.ImageCaptureConfig;
+import androidx.camera.core.ImageProxy;
+import androidx.camera.core.Preview;
+import androidx.camera.core.PreviewConfig;
+import androidx.lifecycle.LifecycleOwner;
+
+public class Camera{
+
+    TextureView textureView;
+    Preview preview;
+    ImageCapture imageCapture;
+    ImageAnalysis imageAnalysis;
+    LifecycleOwner lifecycleOwner;
+
+    public Camera(LifecycleOwner life, TextureView t){
+        textureView = t;
+        lifecycleOwner = life;
+    }
+
+    public void startCamera(ImageCapture ic, ImageAnalysis ia) {
+        CameraX.unbindAll();
+        preview = setPreview();
+        imageCapture = ic;
+        imageAnalysis = ia;
+        //bind to lifecycle:
+        CameraX.bindToLifecycle(lifecycleOwner, preview, imageCapture, imageAnalysis);
+    }
+
+    private Preview setPreview() {
+
+        Rational aspectRatio = new Rational(textureView.getWidth(), textureView.getHeight());
+        Size screen = new Size(textureView.getWidth(), textureView.getHeight()); //size of the screen
+
+
+        PreviewConfig pConfig = new PreviewConfig.Builder().setTargetAspectRatio(aspectRatio).setTargetResolution(screen).build();
+        Preview preview = new Preview(pConfig);
+
+        preview.setOnPreviewOutputUpdateListener(
+                new Preview.OnPreviewOutputUpdateListener() {
+                    @Override
+                    public void onUpdated(Preview.PreviewOutput output) {
+                        ViewGroup parent = (ViewGroup) textureView.getParent();
+                        parent.removeView(textureView);
+                        parent.addView(textureView, 0);
+
+                        textureView.setSurfaceTexture(output.getSurfaceTexture());
+                        updateTransform();
+                    }
+                });
+
+        return preview;
+    }
+
+
+
+
+    private void updateTransform() {
+        Matrix mx = new Matrix();
+        float w = textureView.getMeasuredWidth();
+        float h = textureView.getMeasuredHeight();
+
+        float cX = w / 2f;
+        float cY = h / 2f;
+
+        int rotationDgr;
+        int rotation = (int) textureView.getRotation();
+
+        switch (rotation) {
+            case Surface.ROTATION_0:
+                rotationDgr = 0;
+                break;
+            case Surface.ROTATION_90:
+                rotationDgr = 90;
+                break;
+            case Surface.ROTATION_180:
+                rotationDgr = 180;
+                break;
+            case Surface.ROTATION_270:
+                rotationDgr = 270;
+                break;
+            default:
+                return;
+        }
+
+        mx.postRotate((float) rotationDgr, cX, cY);
+        textureView.setTransform(mx);
+    }
+
+
+
+}
