+++ NotSnap/app/src/main/java/com/COMP3004/NotSnap/SplashActivity.java
@@ -1,0 +1,26 @@
+package com.COMP3004.NotSnap;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+public class SplashActivity extends Activity {
+
+    Handler handler;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.splashscreen);
+
+        handler=new Handler();
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent intent=new Intent(SplashActivity.this,MainActivity.class);
+                startActivity(intent);
+                finish();
+            }
+        },4000);
+
+    }
+}
