+++ NotSnap/app/src/main/res/values/colors.xml
@@ -5,6 +5,7 @@
     <color name="colorAccent">#D81B60</color>
     <color name="rejectedRed">#c62828</color>
     <color name="acceptedGreen">#2e7d32</color>
+    <color name="splash_background_color">#FFFC00</color>
 
 
 </resources>
